[Core][Rebase] Move functions for tx verification
