+++ src/test/transaction_tests.cpp
@@ -8,6 +8,7 @@
 
 #include "clientversion.h"
 #include "consensus/validation.h"
+#include "consensus/tx_verify.h"
 #include "core_io.h"
 #include "key.h"
 #include "keystore.h"
