+++ src/wallet/walletdb.cpp
@@ -9,6 +9,7 @@
 
 #include "base58.h"
 #include "consensus/validation.h"
+#include "consensus/tx_verify.h"
 #include "main.h" // For CheckTransaction
 #include "protocol.h"
 #include "serialize.h"
